Fix actor examples
The Calculator can no more implement the Actor and Server traits
as the former expects the message box to be owned by the actor
while the latter wraps the server with an external message box.
